pyproject.toml
[project]
name = "vbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "pipecat-ai[cartesia,daily,deepgram,google,groq,openai,silero]>=0.0.63",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
